requirements.txt
streamlit
snowflake-connector-python
pandas
requests
gTTS
openai